feat: introduce web bundle
