+++ docs/package.json
@@ -21,6 +21,6 @@
     "unocss": "^0.58.0",
     "unplugin-vue-components": "^0.26.0",
     "vitepress": "1.0.0-rc.30",
-    "vue": "^3.3.11"
+    "vue": "^3.3.13"
   }
 }
