+++ package.json
@@ -24,8 +24,8 @@
     "@rollup/plugin-terser": "^0.4.4",
     "@types/fs-extra": "^11.0.4",
     "@types/hast": "^3.0.3",
-    "@types/node": "^20.10.4",
-    "@vitest/coverage-v8": "^1.0.4",
+    "@types/node": "^20.10.5",
+    "@vitest/coverage-v8": "^1.1.0",
     "ansi-sequence-parser": "^1.1.1",
     "bumpp": "^9.2.1",
     "eslint": "npm:eslint-ts-patch@8.55.0-1",
@@ -42,7 +42,7 @@
     "pnpm": "^8.12.1",
     "prettier": "^3.1.1",
     "rimraf": "^5.0.5",
-    "rollup": "^4.9.0",
+    "rollup": "^4.9.1",
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-dts": "^6.1.0",
     "rollup-plugin-esbuild": "^6.1.0",
@@ -53,9 +53,9 @@
     "typescript": "^5.3.3",
     "unbuild": "^2.0.0",
     "vite": "^5.0.10",
-    "vitest": "^1.0.4",
+    "vitest": "^1.1.0",
     "vue-tsc": "^1.8.25",
-    "wrangler": "^3.21.0"
+    "wrangler": "^3.22.1"
   },
   "pnpm": {
     "patchedDependencies": {
