+++ packages/shikiji-twoslash/package.json
@@ -59,7 +59,7 @@
     "shikiji-core": "workspace:*"
   },
   "devDependencies": {
-    "@iconify-json/carbon": "^1.1.26",
+    "@iconify-json/carbon": "^1.1.27",
     "@iconify-json/codicon": "^1.1.39",
     "hast-util-from-html": "^2.0.1",
     "shiki": "^0.14.7",
