+++ packages/shikiji/scripts/prepare/langs.ts
@@ -22,12 +22,12 @@ export async function prepareLangs() {
     }
 
     const json: LanguageRegistration = {
-      ...lang,
       ...content,
       name: content.name || lang.name,
       scopeName: content.scopeName || lang.scopeName,
       displayName: lang.displayName,
       embeddedLangs: lang.embedded,
+      aliases: lang.aliases,
     }
 
     // F# and Markdown has circular dependency
@@ -43,7 +43,7 @@ import type { LanguageRegistration } from 'shikiji-core'
 
 ${deps.map(i => `import ${i.replace(/[^\w]/g, '_')} from './${i}'`).join('\n')}
 
-const lang = Object.freeze(${JSON.stringify(json, null, 2)}) as unknown as LanguageRegistration
+const lang = Object.freeze(${JSON.stringify(json)}) as unknown as LanguageRegistration
 
 export default [
 ${[
@@ -61,7 +61,7 @@ ${[
       id: i.name,
       name: i.displayName || i.name,
       aliases: i.aliases,
-      import: `__(() => import('./langs/${i.name}')) as DynamicLangReg__`,
+      import: `__(() => import('./langs/${i.name}')) as DynamicImportLanguageRegistration__`,
     }) as const)
       .sort((a, b) => a.id.localeCompare(b.id))
 
@@ -70,16 +70,7 @@ ${[
     await fs.writeFile(
     `src/assets/${fileName}.ts`,
   `${COMMENT_HEAD}
-import type { LanguageRegistration } from 'shikiji-core'
-
-type DynamicLangReg = () => Promise<{ default: LanguageRegistration[] }>
-
-export interface BundledLanguageInfo {
-  id: string
-  name: string
-  import: DynamicLangReg
-  aliases?: string[]
-}
+import type { DynamicImportLanguageRegistration, BundledLanguageInfo } from 'shikiji-core'
 
 export const bundledLanguagesInfo: BundledLanguageInfo[] = ${JSON.stringify(info, null, 2).replace(/"__|__"/g, '').replace(/"/g, '\'')}
 
@@ -87,16 +78,17 @@ export const bundledLanguagesBase = Object.fromEntries(bundledLanguagesInfo.map(
 
 export const bundledLanguagesAlias = Object.fromEntries(bundledLanguagesInfo.flatMap(i => i.aliases?.map(a => [a, i.import]) || []))
 
-export type BuiltinLanguage = ${type}
+export type BundledLanguage = ${type}
 
 export const bundledLanguages = {
   ...bundledLanguagesBase,
   ...bundledLanguagesAlias,
-} as Record<BuiltinLanguage, DynamicLangReg>
+} as Record<BundledLanguage, DynamicImportLanguageRegistration>
 `,
   'utf-8',
     )
   }
 
-  await writeLanguageBundleIndex('langs', grammars.map(i => i.name))
+  await writeLanguageBundleIndex('langs-bundle-full', grammars.map(i => i.name))
+  await writeLanguageBundleIndex('langs-bundle-web', grammars.filter(i => i.categories?.includes('web')).map(i => i.name))
 }
