+++ packages/shikiji/scripts/prepare/themes.ts
@@ -21,28 +21,19 @@ export default Object.freeze(${JSON.stringify(theme, null, 2)}) as unknown as Th
         id: t.name,
         displayName: theme.displayName,
         type: theme.type,
-        import: `__(() => import('./themes/${t.name}')) as unknown as DynamicThemeReg__`,
+        import: `__(() => import('./themes/${t.name}')) as unknown as DynamicImportThemeRegistration__`,
       }
     }))
   await fs.writeFile(
     'src/assets/themes.ts',
 `${COMMENT_HEAD}
-import type { ThemeRegistrationRaw } from 'shikiji-core'
-
-type DynamicThemeReg = () => Promise<{ default: ThemeRegistrationRaw }>
-
-export interface BundledThemeInfo {
-  id: string
-  displayName: string
-  type: 'light' | 'dark'
-  import: DynamicThemeReg
-}
+import type { DynamicImportThemeRegistration, BundledThemeInfo } from 'shikiji-core'
 
 export const bundledThemesInfo: BundledThemeInfo[] = ${JSON.stringify(themes, null, 2).replace(/"__|__"/g, '')}
 
-export type BuiltinTheme = ${themes.map(i => `'${i.id}'`).join(' | ')}
+export type BundledTheme = ${themes.map(i => `'${i.id}'`).join(' | ')}
 
-export const bundledThemes = Object.fromEntries(bundledThemesInfo.map(i => [i.id, i.import])) as Record<BuiltinTheme, DynamicThemeReg>
+export const bundledThemes = Object.fromEntries(bundledThemesInfo.map(i => [i.id, i.import])) as Record<BundledTheme, DynamicImportThemeRegistration>
 `,
 'utf-8',
   )
