+++ packages/shikiji/src/assets/langs-bundle-full.ts
@@ -1,0 +1,1029 @@
+/**
+ * Generated by scripts/prepare.ts
+ */
+import type { DynamicImportLanguageRegistration, BundledLanguageInfo } from 'shikiji-core'
+
+export const bundledLanguagesInfo: BundledLanguageInfo[] = [
+  {
+    'id': 'abap',
+    'name': 'ABAP',
+    'import': (() => import('./langs/abap')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'actionscript-3',
+    'name': 'ActionScript',
+    'import': (() => import('./langs/actionscript-3')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ada',
+    'name': 'Ada',
+    'import': (() => import('./langs/ada')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'apache',
+    'name': 'Apache Conf',
+    'import': (() => import('./langs/apache')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'apex',
+    'name': 'Apex',
+    'import': (() => import('./langs/apex')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'apl',
+    'name': 'APL',
+    'import': (() => import('./langs/apl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'applescript',
+    'name': 'AppleScript',
+    'import': (() => import('./langs/applescript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ara',
+    'name': 'Ara',
+    'import': (() => import('./langs/ara')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'asm',
+    'name': 'Assembly',
+    'import': (() => import('./langs/asm')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'astro',
+    'name': 'Astro',
+    'import': (() => import('./langs/astro')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'awk',
+    'name': 'AWK',
+    'import': (() => import('./langs/awk')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ballerina',
+    'name': 'Ballerina',
+    'import': (() => import('./langs/ballerina')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'bat',
+    'name': 'Batch File',
+    'aliases': [
+      'batch'
+    ],
+    'import': (() => import('./langs/bat')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'beancount',
+    'name': 'Beancount',
+    'import': (() => import('./langs/beancount')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'berry',
+    'name': 'Berry',
+    'aliases': [
+      'be'
+    ],
+    'import': (() => import('./langs/berry')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'bibtex',
+    'name': 'BibTeX',
+    'import': (() => import('./langs/bibtex')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'bicep',
+    'name': 'Bicep',
+    'import': (() => import('./langs/bicep')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'blade',
+    'name': 'Blade',
+    'import': (() => import('./langs/blade')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'c',
+    'name': 'C',
+    'import': (() => import('./langs/c')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cadence',
+    'name': 'Cadence',
+    'aliases': [
+      'cdc'
+    ],
+    'import': (() => import('./langs/cadence')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'clarity',
+    'name': 'Clarity',
+    'import': (() => import('./langs/clarity')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'clojure',
+    'name': 'Clojure',
+    'aliases': [
+      'clj'
+    ],
+    'import': (() => import('./langs/clojure')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cmake',
+    'name': 'CMake',
+    'import': (() => import('./langs/cmake')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cobol',
+    'name': 'COBOL',
+    'import': (() => import('./langs/cobol')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'codeql',
+    'name': 'CodeQL',
+    'aliases': [
+      'ql'
+    ],
+    'import': (() => import('./langs/codeql')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'coffee',
+    'name': 'CoffeeScript',
+    'aliases': [
+      'coffeescript'
+    ],
+    'import': (() => import('./langs/coffee')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cpp',
+    'name': 'C++',
+    'aliases': [
+      'c++'
+    ],
+    'import': (() => import('./langs/cpp')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'crystal',
+    'name': 'Crystal',
+    'import': (() => import('./langs/crystal')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'csharp',
+    'name': 'C#',
+    'aliases': [
+      'c#',
+      'cs'
+    ],
+    'import': (() => import('./langs/csharp')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'css',
+    'name': 'CSS',
+    'import': (() => import('./langs/css')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'csv',
+    'name': 'csv syntax',
+    'import': (() => import('./langs/csv')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cue',
+    'name': 'CUE',
+    'import': (() => import('./langs/cue')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'cypher',
+    'name': 'Cypher',
+    'aliases': [
+      'cql'
+    ],
+    'import': (() => import('./langs/cypher')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'd',
+    'name': 'D',
+    'import': (() => import('./langs/d')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'dart',
+    'name': 'Dart',
+    'import': (() => import('./langs/dart')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'dax',
+    'name': 'DAX',
+    'import': (() => import('./langs/dax')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'diff',
+    'name': 'Diff',
+    'import': (() => import('./langs/diff')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'docker',
+    'name': 'Dockerfile',
+    'aliases': [
+      'dockerfile'
+    ],
+    'import': (() => import('./langs/docker')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'dream-maker',
+    'name': 'Dream Maker',
+    'import': (() => import('./langs/dream-maker')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'elixir',
+    'name': 'Elixir',
+    'import': (() => import('./langs/elixir')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'elm',
+    'name': 'Elm',
+    'import': (() => import('./langs/elm')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'erb',
+    'name': 'ERB',
+    'import': (() => import('./langs/erb')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'erlang',
+    'name': 'Erlang',
+    'aliases': [
+      'erl'
+    ],
+    'import': (() => import('./langs/erlang')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'fish',
+    'name': 'Fish',
+    'import': (() => import('./langs/fish')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'fsharp',
+    'name': 'F#',
+    'aliases': [
+      'f#',
+      'fs'
+    ],
+    'import': (() => import('./langs/fsharp')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'gdresource',
+    'name': 'GDResource',
+    'import': (() => import('./langs/gdresource')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'gdscript',
+    'name': 'GDScript',
+    'import': (() => import('./langs/gdscript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'gdshader',
+    'name': 'GDShader',
+    'import': (() => import('./langs/gdshader')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'gherkin',
+    'name': 'Gherkin',
+    'import': (() => import('./langs/gherkin')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'git-commit',
+    'name': 'Git Commit Message',
+    'import': (() => import('./langs/git-commit')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'git-rebase',
+    'name': 'Git Rebase Message',
+    'import': (() => import('./langs/git-rebase')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'glimmer-js',
+    'name': 'Glimmer JS',
+    'aliases': [
+      'gjs'
+    ],
+    'import': (() => import('./langs/glimmer-js')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'glimmer-ts',
+    'name': 'Glimmer TS',
+    'aliases': [
+      'gts'
+    ],
+    'import': (() => import('./langs/glimmer-ts')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'glsl',
+    'name': 'GLSL',
+    'import': (() => import('./langs/glsl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'gnuplot',
+    'name': 'Gnuplot',
+    'import': (() => import('./langs/gnuplot')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'go',
+    'name': 'Go',
+    'import': (() => import('./langs/go')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'graphql',
+    'name': 'GraphQL',
+    'aliases': [
+      'gql'
+    ],
+    'import': (() => import('./langs/graphql')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'groovy',
+    'name': 'Groovy',
+    'import': (() => import('./langs/groovy')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'hack',
+    'name': 'Hack',
+    'import': (() => import('./langs/hack')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'haml',
+    'name': 'Ruby Haml',
+    'import': (() => import('./langs/haml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'handlebars',
+    'name': 'Handlebars',
+    'aliases': [
+      'hbs'
+    ],
+    'import': (() => import('./langs/handlebars')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'haskell',
+    'name': 'Haskell',
+    'aliases': [
+      'hs'
+    ],
+    'import': (() => import('./langs/haskell')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'hcl',
+    'name': 'HashiCorp HCL',
+    'import': (() => import('./langs/hcl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'hjson',
+    'name': 'Hjson',
+    'import': (() => import('./langs/hjson')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'hlsl',
+    'name': 'HLSL',
+    'import': (() => import('./langs/hlsl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'html',
+    'name': 'HTML',
+    'import': (() => import('./langs/html')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'http',
+    'name': 'HTTP',
+    'import': (() => import('./langs/http')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'imba',
+    'name': 'Imba',
+    'import': (() => import('./langs/imba')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ini',
+    'name': 'INI',
+    'aliases': [
+      'properties'
+    ],
+    'import': (() => import('./langs/ini')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'java',
+    'name': 'Java',
+    'import': (() => import('./langs/java')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'javascript',
+    'name': 'JavaScript',
+    'aliases': [
+      'js'
+    ],
+    'import': (() => import('./langs/javascript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jinja',
+    'name': 'Jinja',
+    'import': (() => import('./langs/jinja')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jison',
+    'name': 'Jison',
+    'import': (() => import('./langs/jison')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'json',
+    'name': 'JSON',
+    'import': (() => import('./langs/json')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'json5',
+    'name': 'JSON5',
+    'import': (() => import('./langs/json5')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jsonc',
+    'name': 'JSON with Comments',
+    'import': (() => import('./langs/jsonc')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jsonl',
+    'name': 'JSON Lines',
+    'import': (() => import('./langs/jsonl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jsonnet',
+    'name': 'Jsonnet',
+    'import': (() => import('./langs/jsonnet')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jssm',
+    'name': 'JSSM',
+    'aliases': [
+      'fsl'
+    ],
+    'import': (() => import('./langs/jssm')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'jsx',
+    'name': 'JSX',
+    'import': (() => import('./langs/jsx')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'julia',
+    'name': 'Julia',
+    'import': (() => import('./langs/julia')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'kotlin',
+    'name': 'Kotlin',
+    'aliases': [
+      'kt',
+      'kts'
+    ],
+    'import': (() => import('./langs/kotlin')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'kusto',
+    'name': 'Kusto',
+    'aliases': [
+      'kql'
+    ],
+    'import': (() => import('./langs/kusto')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'latex',
+    'name': 'LaTeX',
+    'import': (() => import('./langs/latex')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'less',
+    'name': 'Less',
+    'import': (() => import('./langs/less')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'liquid',
+    'name': 'Liquid',
+    'import': (() => import('./langs/liquid')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'lisp',
+    'name': 'Lisp',
+    'import': (() => import('./langs/lisp')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'logo',
+    'name': 'Logo',
+    'import': (() => import('./langs/logo')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'lua',
+    'name': 'Lua',
+    'import': (() => import('./langs/lua')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'make',
+    'name': 'Makefile',
+    'aliases': [
+      'makefile'
+    ],
+    'import': (() => import('./langs/make')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'markdown',
+    'name': 'Markdown',
+    'aliases': [
+      'md'
+    ],
+    'import': (() => import('./langs/markdown')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'marko',
+    'name': 'Marko',
+    'import': (() => import('./langs/marko')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'matlab',
+    'name': 'MATLAB',
+    'import': (() => import('./langs/matlab')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'mdc',
+    'name': 'mdc',
+    'import': (() => import('./langs/mdc')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'mdx',
+    'name': 'MDX',
+    'import': (() => import('./langs/mdx')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'mermaid',
+    'name': 'Mermaid',
+    'import': (() => import('./langs/mermaid')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'mojo',
+    'name': 'Mojo',
+    'import': (() => import('./langs/mojo')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'narrat',
+    'name': 'Narrat Language',
+    'aliases': [
+      'nar'
+    ],
+    'import': (() => import('./langs/narrat')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'nextflow',
+    'name': 'Nextflow',
+    'aliases': [
+      'nf'
+    ],
+    'import': (() => import('./langs/nextflow')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'nginx',
+    'name': 'Nginx',
+    'import': (() => import('./langs/nginx')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'nim',
+    'name': 'Nim',
+    'import': (() => import('./langs/nim')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'nix',
+    'name': 'Nix',
+    'import': (() => import('./langs/nix')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'nushell',
+    'name': 'nushell',
+    'aliases': [
+      'nu'
+    ],
+    'import': (() => import('./langs/nushell')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'objective-c',
+    'name': 'Objective-C',
+    'aliases': [
+      'objc'
+    ],
+    'import': (() => import('./langs/objective-c')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'objective-cpp',
+    'name': 'Objective-C++',
+    'import': (() => import('./langs/objective-cpp')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ocaml',
+    'name': 'OCaml',
+    'import': (() => import('./langs/ocaml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'pascal',
+    'name': 'Pascal',
+    'import': (() => import('./langs/pascal')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'perl',
+    'name': 'Perl',
+    'import': (() => import('./langs/perl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'php',
+    'name': 'PHP',
+    'import': (() => import('./langs/php')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'plsql',
+    'name': 'PL/SQL',
+    'import': (() => import('./langs/plsql')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'postcss',
+    'name': 'PostCSS',
+    'import': (() => import('./langs/postcss')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'powerquery',
+    'name': 'PowerQuery',
+    'import': (() => import('./langs/powerquery')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'powershell',
+    'name': 'PowerShell',
+    'aliases': [
+      'ps',
+      'ps1'
+    ],
+    'import': (() => import('./langs/powershell')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'prisma',
+    'name': 'Prisma',
+    'import': (() => import('./langs/prisma')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'prolog',
+    'name': 'Prolog',
+    'import': (() => import('./langs/prolog')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'proto',
+    'name': 'Protocol Buffer 3',
+    'import': (() => import('./langs/proto')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'pug',
+    'name': 'Pug',
+    'aliases': [
+      'jade'
+    ],
+    'import': (() => import('./langs/pug')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'puppet',
+    'name': 'Puppet',
+    'import': (() => import('./langs/puppet')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'purescript',
+    'name': 'PureScript',
+    'import': (() => import('./langs/purescript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'python',
+    'name': 'Python',
+    'aliases': [
+      'py'
+    ],
+    'import': (() => import('./langs/python')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'r',
+    'name': 'R',
+    'import': (() => import('./langs/r')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'raku',
+    'name': 'Raku',
+    'aliases': [
+      'perl6'
+    ],
+    'import': (() => import('./langs/raku')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'razor',
+    'name': 'ASP.NET Razor',
+    'import': (() => import('./langs/razor')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'reg',
+    'name': 'Windows Registry Script',
+    'import': (() => import('./langs/reg')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'rel',
+    'name': 'Rel',
+    'import': (() => import('./langs/rel')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'riscv',
+    'name': 'RISC-V',
+    'import': (() => import('./langs/riscv')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'rst',
+    'name': 'reStructuredText',
+    'import': (() => import('./langs/rst')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ruby',
+    'name': 'Ruby',
+    'aliases': [
+      'rb'
+    ],
+    'import': (() => import('./langs/ruby')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'rust',
+    'name': 'Rust',
+    'aliases': [
+      'rs'
+    ],
+    'import': (() => import('./langs/rust')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'sas',
+    'name': 'SAS',
+    'import': (() => import('./langs/sas')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'sass',
+    'name': 'Sass',
+    'import': (() => import('./langs/sass')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'scala',
+    'name': 'Scala',
+    'import': (() => import('./langs/scala')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'scheme',
+    'name': 'Scheme',
+    'import': (() => import('./langs/scheme')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'scss',
+    'name': 'SCSS',
+    'import': (() => import('./langs/scss')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'shaderlab',
+    'name': 'ShaderLab',
+    'aliases': [
+      'shader'
+    ],
+    'import': (() => import('./langs/shaderlab')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'shellscript',
+    'name': 'Shell',
+    'aliases': [
+      'bash',
+      'sh',
+      'shell',
+      'zsh'
+    ],
+    'import': (() => import('./langs/shellscript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'shellsession',
+    'name': 'Shell Session',
+    'aliases': [
+      'console'
+    ],
+    'import': (() => import('./langs/shellsession')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'smalltalk',
+    'name': 'Smalltalk',
+    'import': (() => import('./langs/smalltalk')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'solidity',
+    'name': 'Solidity',
+    'import': (() => import('./langs/solidity')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'sparql',
+    'name': 'SPARQL',
+    'import': (() => import('./langs/sparql')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'splunk',
+    'name': 'Splunk Query Language',
+    'aliases': [
+      'spl'
+    ],
+    'import': (() => import('./langs/splunk')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'sql',
+    'name': 'SQL',
+    'import': (() => import('./langs/sql')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'ssh-config',
+    'name': 'SSH Config',
+    'import': (() => import('./langs/ssh-config')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'stata',
+    'name': 'Stata',
+    'import': (() => import('./langs/stata')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'stylus',
+    'name': 'Stylus',
+    'aliases': [
+      'styl'
+    ],
+    'import': (() => import('./langs/stylus')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'svelte',
+    'name': 'Svelte',
+    'import': (() => import('./langs/svelte')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'swift',
+    'name': 'Swift',
+    'import': (() => import('./langs/swift')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'system-verilog',
+    'name': 'SystemVerilog',
+    'import': (() => import('./langs/system-verilog')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'tasl',
+    'name': 'Tasl',
+    'import': (() => import('./langs/tasl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'tcl',
+    'name': 'Tcl',
+    'import': (() => import('./langs/tcl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'tex',
+    'name': 'TeX',
+    'import': (() => import('./langs/tex')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'toml',
+    'name': 'TOML',
+    'import': (() => import('./langs/toml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'tsx',
+    'name': 'TSX',
+    'import': (() => import('./langs/tsx')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'turtle',
+    'name': 'Turtle',
+    'import': (() => import('./langs/turtle')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'twig',
+    'name': 'Twig',
+    'import': (() => import('./langs/twig')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'typescript',
+    'name': 'TypeScript',
+    'aliases': [
+      'ts'
+    ],
+    'import': (() => import('./langs/typescript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'v',
+    'name': 'V',
+    'import': (() => import('./langs/v')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'vb',
+    'name': 'Visual Basic',
+    'aliases': [
+      'cmd'
+    ],
+    'import': (() => import('./langs/vb')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'verilog',
+    'name': 'Verilog',
+    'import': (() => import('./langs/verilog')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'vhdl',
+    'name': 'VHDL',
+    'import': (() => import('./langs/vhdl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'viml',
+    'name': 'Vim Script',
+    'aliases': [
+      'vim',
+      'vimscript'
+    ],
+    'import': (() => import('./langs/viml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'vue',
+    'name': 'Vue',
+    'import': (() => import('./langs/vue')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'vue-html',
+    'name': 'Vue HTML',
+    'import': (() => import('./langs/vue-html')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'vyper',
+    'name': 'Vyper',
+    'aliases': [
+      'vy'
+    ],
+    'import': (() => import('./langs/vyper')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'wasm',
+    'name': 'WebAssembly',
+    'import': (() => import('./langs/wasm')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'wenyan',
+    'name': 'Wenyan',
+    'aliases': [
+      '文言'
+    ],
+    'import': (() => import('./langs/wenyan')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'wgsl',
+    'name': 'WGSL',
+    'import': (() => import('./langs/wgsl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'wolfram',
+    'name': 'Wolfram',
+    'aliases': [
+      'wl'
+    ],
+    'import': (() => import('./langs/wolfram')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'xml',
+    'name': 'XML',
+    'import': (() => import('./langs/xml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'xsl',
+    'name': 'XSL',
+    'import': (() => import('./langs/xsl')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'yaml',
+    'name': 'YAML',
+    'aliases': [
+      'yml'
+    ],
+    'import': (() => import('./langs/yaml')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'zenscript',
+    'name': 'ZenScript',
+    'import': (() => import('./langs/zenscript')) as DynamicImportLanguageRegistration
+  },
+  {
+    'id': 'zig',
+    'name': 'zig',
+    'import': (() => import('./langs/zig')) as DynamicImportLanguageRegistration
+  }
+]
+
+export const bundledLanguagesBase = Object.fromEntries(bundledLanguagesInfo.map(i => [i.id, i.import]))
+
+export const bundledLanguagesAlias = Object.fromEntries(bundledLanguagesInfo.flatMap(i => i.aliases?.map(a => [a, i.import]) || []))
+
+export type BundledLanguage = 'abap' | 'actionscript-3' | 'ada' | 'apache' | 'apex' | 'apl' | 'applescript' | 'ara' | 'asm' | 'astro' | 'awk' | 'ballerina' | 'bash' | 'bat' | 'batch' | 'be' | 'beancount' | 'berry' | 'bibtex' | 'bicep' | 'blade' | 'c' | 'c#' | 'c++' | 'cadence' | 'cdc' | 'clarity' | 'clj' | 'clojure' | 'cmake' | 'cmd' | 'cobol' | 'codeql' | 'coffee' | 'coffeescript' | 'console' | 'cpp' | 'cql' | 'crystal' | 'cs' | 'csharp' | 'css' | 'csv' | 'cue' | 'cypher' | 'd' | 'dart' | 'dax' | 'diff' | 'docker' | 'dockerfile' | 'dream-maker' | 'elixir' | 'elm' | 'erb' | 'erl' | 'erlang' | 'f#' | 'fish' | 'fs' | 'fsharp' | 'fsl' | 'gdresource' | 'gdscript' | 'gdshader' | 'gherkin' | 'git-commit' | 'git-rebase' | 'gjs' | 'glimmer-js' | 'glimmer-ts' | 'glsl' | 'gnuplot' | 'go' | 'gql' | 'graphql' | 'groovy' | 'gts' | 'hack' | 'haml' | 'handlebars' | 'haskell' | 'hbs' | 'hcl' | 'hjson' | 'hlsl' | 'hs' | 'html' | 'http' | 'imba' | 'ini' | 'jade' | 'java' | 'javascript' | 'jinja' | 'jison' | 'js' | 'json' | 'json5' | 'jsonc' | 'jsonl' | 'jsonnet' | 'jssm' | 'jsx' | 'julia' | 'kotlin' | 'kql' | 'kt' | 'kts' | 'kusto' | 'latex' | 'less' | 'liquid' | 'lisp' | 'logo' | 'lua' | 'make' | 'makefile' | 'markdown' | 'marko' | 'matlab' | 'md' | 'mdc' | 'mdx' | 'mermaid' | 'mojo' | 'nar' | 'narrat' | 'nextflow' | 'nf' | 'nginx' | 'nim' | 'nix' | 'nu' | 'nushell' | 'objc' | 'objective-c' | 'objective-cpp' | 'ocaml' | 'pascal' | 'perl' | 'perl6' | 'php' | 'plsql' | 'postcss' | 'powerquery' | 'powershell' | 'prisma' | 'prolog' | 'properties' | 'proto' | 'ps' | 'ps1' | 'pug' | 'puppet' | 'purescript' | 'py' | 'python' | 'ql' | 'r' | 'raku' | 'razor' | 'rb' | 'reg' | 'rel' | 'riscv' | 'rs' | 'rst' | 'ruby' | 'rust' | 'sas' | 'sass' | 'scala' | 'scheme' | 'scss' | 'sh' | 'shader' | 'shaderlab' | 'shell' | 'shellscript' | 'shellsession' | 'smalltalk' | 'solidity' | 'sparql' | 'spl' | 'splunk' | 'sql' | 'ssh-config' | 'stata' | 'styl' | 'stylus' | 'svelte' | 'swift' | 'system-verilog' | 'tasl' | 'tcl' | 'tex' | 'toml' | 'ts' | 'tsx' | 'turtle' | 'twig' | 'typescript' | 'v' | 'vb' | 'verilog' | 'vhdl' | 'vim' | 'viml' | 'vimscript' | 'vue' | 'vue-html' | 'vy' | 'vyper' | 'wasm' | 'wenyan' | 'wgsl' | 'wl' | 'wolfram' | 'xml' | 'xsl' | 'yaml' | 'yml' | 'zenscript' | 'zig' | 'zsh' | '文言'
+
+export const bundledLanguages = {
+  ...bundledLanguagesBase,
+  ...bundledLanguagesAlias,
+} as Record<BundledLanguage, DynamicImportLanguageRegistration>
