+++ packages/shikiji/src/assets/themes.ts
@@ -1,188 +1,179 @@
 /**
  * Generated by scripts/prepare.ts
  */
-import type { ThemeRegistrationRaw } from 'shikiji-core'
-
-type DynamicThemeReg = () => Promise<{ default: ThemeRegistrationRaw }>
-
-export interface BundledThemeInfo {
-  id: string
-  displayName: string
-  type: 'light' | 'dark'
-  import: DynamicThemeReg
-}
+import type { DynamicImportThemeRegistration, BundledThemeInfo } from 'shikiji-core'
 
 export const bundledThemesInfo: BundledThemeInfo[] = [
   {
     "id": "dark-plus",
     "displayName": "Dark Plus",
     "type": "dark",
-    "import": (() => import('./themes/dark-plus')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/dark-plus')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "dracula",
     "displayName": "Dracula",
     "type": "dark",
-    "import": (() => import('./themes/dracula')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/dracula')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "dracula-soft",
     "displayName": "Dracula Soft",
     "type": "dark",
-    "import": (() => import('./themes/dracula-soft')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/dracula-soft')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "github-dark",
     "displayName": "GitHub Dark",
     "type": "dark",
-    "import": (() => import('./themes/github-dark')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/github-dark')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "github-dark-dimmed",
     "displayName": "GitHub Dark Dimmed",
     "type": "dark",
-    "import": (() => import('./themes/github-dark-dimmed')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/github-dark-dimmed')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "github-light",
     "displayName": "GitHub Light",
     "type": "light",
-    "import": (() => import('./themes/github-light')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/github-light')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "light-plus",
     "displayName": "Light Plus",
     "type": "light",
-    "import": (() => import('./themes/light-plus')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/light-plus')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "material-theme",
     "displayName": "Material Theme",
     "type": "dark",
-    "import": (() => import('./themes/material-theme')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/material-theme')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "material-theme-darker",
     "displayName": "Material Theme Darker",
     "type": "dark",
-    "import": (() => import('./themes/material-theme-darker')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/material-theme-darker')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "material-theme-lighter",
     "displayName": "Material Theme Lighter",
     "type": "light",
-    "import": (() => import('./themes/material-theme-lighter')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/material-theme-lighter')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "material-theme-ocean",
     "displayName": "Material Theme Ocean",
     "type": "dark",
-    "import": (() => import('./themes/material-theme-ocean')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/material-theme-ocean')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "material-theme-palenight",
     "displayName": "Material Theme Palenight",
     "type": "dark",
-    "import": (() => import('./themes/material-theme-palenight')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/material-theme-palenight')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "min-dark",
     "displayName": "Min Dark",
     "type": "dark",
-    "import": (() => import('./themes/min-dark')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/min-dark')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "min-light",
     "displayName": "Min Light",
     "type": "light",
-    "import": (() => import('./themes/min-light')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/min-light')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "monokai",
     "displayName": "Monokai",
     "type": "dark",
-    "import": (() => import('./themes/monokai')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/monokai')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "nord",
     "displayName": "Nord",
     "type": "dark",
-    "import": (() => import('./themes/nord')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/nord')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "one-dark-pro",
     "displayName": "One Dark Pro",
     "type": "dark",
-    "import": (() => import('./themes/one-dark-pro')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/one-dark-pro')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "poimandres",
     "displayName": "Poimandres",
     "type": "dark",
-    "import": (() => import('./themes/poimandres')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/poimandres')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "rose-pine",
     "displayName": "Rosé Pine",
     "type": "dark",
-    "import": (() => import('./themes/rose-pine')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/rose-pine')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "rose-pine-dawn",
     "displayName": "Rosé Pine Dawn",
     "type": "light",
-    "import": (() => import('./themes/rose-pine-dawn')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/rose-pine-dawn')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "rose-pine-moon",
     "displayName": "Rosé Pine Moon",
     "type": "dark",
-    "import": (() => import('./themes/rose-pine-moon')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/rose-pine-moon')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "slack-dark",
     "displayName": "Slack Dark",
     "type": "dark",
-    "import": (() => import('./themes/slack-dark')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/slack-dark')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "slack-ochin",
     "displayName": "Slack Ochin",
     "type": "light",
-    "import": (() => import('./themes/slack-ochin')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/slack-ochin')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "solarized-dark",
     "displayName": "Solarized Dark",
     "type": "dark",
-    "import": (() => import('./themes/solarized-dark')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/solarized-dark')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "solarized-light",
     "displayName": "Solarized Light",
     "type": "light",
-    "import": (() => import('./themes/solarized-light')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/solarized-light')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "vitesse-black",
     "displayName": "Vitesse Black",
     "type": "dark",
-    "import": (() => import('./themes/vitesse-black')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/vitesse-black')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "vitesse-dark",
     "displayName": "Vitesse Dark",
     "type": "dark",
-    "import": (() => import('./themes/vitesse-dark')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/vitesse-dark')) as unknown as DynamicImportThemeRegistration
   },
   {
     "id": "vitesse-light",
     "displayName": "Vitesse Light",
     "type": "light",
-    "import": (() => import('./themes/vitesse-light')) as unknown as DynamicThemeReg
+    "import": (() => import('./themes/vitesse-light')) as unknown as DynamicImportThemeRegistration
   }
 ]
 
-export type BuiltinTheme = 'dark-plus' | 'dracula' | 'dracula-soft' | 'github-dark' | 'github-dark-dimmed' | 'github-light' | 'light-plus' | 'material-theme' | 'material-theme-darker' | 'material-theme-lighter' | 'material-theme-ocean' | 'material-theme-palenight' | 'min-dark' | 'min-light' | 'monokai' | 'nord' | 'one-dark-pro' | 'poimandres' | 'rose-pine' | 'rose-pine-dawn' | 'rose-pine-moon' | 'slack-dark' | 'slack-ochin' | 'solarized-dark' | 'solarized-light' | 'vitesse-black' | 'vitesse-dark' | 'vitesse-light'
+export type BundledTheme = 'dark-plus' | 'dracula' | 'dracula-soft' | 'github-dark' | 'github-dark-dimmed' | 'github-light' | 'light-plus' | 'material-theme' | 'material-theme-darker' | 'material-theme-lighter' | 'material-theme-ocean' | 'material-theme-palenight' | 'min-dark' | 'min-light' | 'monokai' | 'nord' | 'one-dark-pro' | 'poimandres' | 'rose-pine' | 'rose-pine-dawn' | 'rose-pine-moon' | 'slack-dark' | 'slack-ochin' | 'solarized-dark' | 'solarized-light' | 'vitesse-black' | 'vitesse-dark' | 'vitesse-light'
 
-export const bundledThemes = Object.fromEntries(bundledThemesInfo.map(i => [i.id, i.import])) as Record<BuiltinTheme, DynamicThemeReg>
+export const bundledThemes = Object.fromEntries(bundledThemesInfo.map(i => [i.id, i.import])) as Record<BundledTheme, DynamicImportThemeRegistration>
