+++ packages/shikiji/src/bundle-full.ts
@@ -1,0 +1,36 @@
+import type { HighlighterGeneric } from 'shikiji-core'
+import { createSingletonShorthands, createdBundledHighlighter } from './core'
+import type { BundledLanguage } from './assets/langs-bundle-full'
+import type { BundledTheme } from './themes'
+import { bundledLanguages } from './assets/langs-bundle-full'
+import { bundledThemes } from './themes'
+import { getWasmInlined } from './wasm'
+
+export * from './core'
+export * from './themes'
+export * from './assets/langs-bundle-full'
+export * from './wasm'
+
+export type Highlighter = HighlighterGeneric<BundledLanguage, BundledTheme>
+
+export const getHighlighter = /* @__PURE__ */ createdBundledHighlighter<
+  BundledLanguage,
+  BundledTheme
+>(
+  bundledLanguages,
+  bundledThemes,
+  getWasmInlined,
+)
+
+export const {
+  codeToHtml,
+  codeToHast,
+  codeToThemedTokens,
+  codeToTokensWithThemes,
+  getSingletonHighlighter,
+} = /* @__PURE__ */ createSingletonShorthands<
+  BundledLanguage,
+  BundledTheme
+>(
+  getHighlighter,
+)
