+++ packages/shikiji/src/types.ts
@@ -1,4 +1,10 @@
-export type { BuiltinLanguage, BundledLanguageInfo } from './langs'
-export type { BuiltinTheme, BundledThemeInfo } from './themes'
+import type { BundledLanguage } from './langs'
+import type { BundledTheme } from './themes'
+
+export type { BundledLanguage } from './langs'
+export type { BundledTheme } from './themes'
 
 export type * from 'shikiji-core/types'
+
+export type BuiltinLanguage = BundledLanguage
+export type BuiltinTheme = BundledTheme
